backend/app/requirements.txt
fastapi
uvicorn[standard]
mysql-connector-python
python-dotenv
openai-whisper
python-multipart
torch
transformers
mistralai
